+++ src/Products/PloneMeeting/tests/PloneMeetingTestCase.py
@@ -204,6 +204,9 @@ class PloneMeetingTestCase(unittest.TestCase, PloneMeetingTestingHelpers):
         '''Logs out currently logged user and logs in p_loginName.'''
         logout()
         self.cleanMemoize()
+        cleanRamCacheFor('Products.PloneMeeting.ToolPloneMeeting._users_groups_value')
+        cleanRamCacheFor('Products.PloneMeeting.ToolPloneMeeting.get_plone_groups_for_user')
+        cleanRamCacheFor('Products.PloneMeeting.ToolPloneMeeting.isPowerObserverForCfg')
         if loginName == 'admin':
             login(self.app, loginName)
         else:
@@ -535,7 +538,8 @@ class PloneMeetingTestCase(unittest.TestCase, PloneMeetingTestingHelpers):
                                 cfg=None,
                                 observer_type='powerobservers',
                                 field_name='item_states',
-                                states=[]):
+                                states=[],
+                                access_on=''):
         """Change power observers states for item or meeting."""
         if not cfg:
             cfg = self.meetingConfig
@@ -543,4 +547,8 @@ class PloneMeetingTestCase(unittest.TestCase, PloneMeetingTestingHelpers):
         for po_infos in power_observers:
             if po_infos['row_id'] == observer_type:
                 po_infos[field_name] = states
+                if field_name == 'item_states':
+                    po_infos['item_access_on'] = access_on
+                else:
+                    po_infos['meeting_access_on'] = access_on
         cfg.setPowerObservers(power_observers)