+++ src/Products/PloneMeeting/tests/testToolPloneMeeting.py
@@ -747,9 +747,13 @@ class testToolPloneMeeting(PloneMeetingTestCase):
            First set (restricted) power observers may view in state 'itemcreated' then change to 'proposed'."""
         cfg = self.meetingConfig
         self._setPowerObserverStates(states=('itemcreated', ))
-        cfg.setMeetingPowerObserversStates(('created', ))
-        cfg.setItemRestrictedPowerObserversStates((self._stateMappingFor('proposed'), ))
-        cfg.setMeetingRestrictedPowerObserversStates(('closed', ))
+        self._setPowerObserverStates(field_name='meeting_states',
+                                     states=('created', ))
+        self._setPowerObserverStates(observer_type='restrictedpowerobservers',
+                                     states=(self._stateMappingFor('proposed'), ))
+        self._setPowerObserverStates(field_name='meeting_states',
+                                     observer_type='restrictedpowerobservers',
+                                     states=('closed', ))
         # only available to 'Managers'
         self.changeUser('pmManager')
         self.assertRaises(Unauthorized, self.tool.restrictedTraverse, 'updateAllLocalRoles')
@@ -769,9 +773,13 @@ class testToolPloneMeeting(PloneMeetingTestCase):
         # change configuration, updateAllLocalRoles then check again
         self.changeUser('siteadmin')
         self._setPowerObserverStates(states=(self._stateMappingFor('proposed'), ))
-        cfg.setMeetingPowerObserversStates(('closed', ))
-        cfg.setItemRestrictedPowerObserversStates(('itemcreated', ))
-        cfg.setMeetingRestrictedPowerObserversStates(('created', ))
+        self._setPowerObserverStates(field_name='meeting_states',
+                                     states=('closed', ))
+        self._setPowerObserverStates(observer_type='restrictedpowerobservers',
+                                     states=('itemcreated', ))
+        self._setPowerObserverStates(field_name='meeting_states',
+                                     observer_type='restrictedpowerobservers',
+                                     states=('created', ))
         self.tool.updateAllLocalRoles()
         # local roles and catalog are updated
         catalog = self.portal.portal_catalog
@@ -1031,7 +1039,23 @@ class testToolPloneMeeting(PloneMeetingTestCase):
 
     def test_pm_IsPowerObserverForCfg(self):
         """ """
-        import ipdb; ipdb.set_trace()
+        cfg = self.meetingConfig
+        self.changeUser('pmManager')
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='powerobservers'))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='restrictedpowerobservers'))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='unknown'))
+        self.changeUser('powerobserver1')
+        self.assertTrue(self.tool.isPowerObserverForCfg(cfg))
+        self.assertTrue(self.tool.isPowerObserverForCfg(cfg, power_observer_type='powerobservers'))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='restrictedpowerobservers'))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='unknown'))
+        self.changeUser('restrictedpowerobserver1')
+        self.assertTrue(self.tool.isPowerObserverForCfg(cfg))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='powerobservers'))
+        self.assertTrue(self.tool.isPowerObserverForCfg(cfg, power_observer_type='restrictedpowerobservers'))
+        self.assertFalse(self.tool.isPowerObserverForCfg(cfg, power_observer_type='unknown'))
+
 
 def test_suite():
     from unittest import TestSuite, makeSuite